requirements/requirements-optionals.txt
# Optional packages which may be used with REST framework.
psycopg2-binary>=2.8.2, <2.9
markdown==3.1.1
pygments==2.4.2
django-guardian==1.5.0
django-filter>=2.1.0, <2.2
coreapi==2.3.1
coreschema==0.0.4
pyyaml>=5.1
requests>=2.22.0